refactor: move LayerMapper and BiomeMapper to `world/generator`
